refactor: split platform and app layouts
